refactor: rename NewPokemonDetailSkillAdapter to PokemonDetailSkillAdapter
